+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 * Add CanBuildOutputAst function to RevitDropDownBase for judging whether it have valid Enumeration values to the output in dropdown list
+* Update RevitSystemTestBase to make it more user friendly
 
 ## 0.1.0
 * The version of master branch will be start with 0.1.x, and Revit2020 branch will still be 0.0.xx.